Rename every import by adding .js extension as per https://google.github.io/styleguide/jsguide.html#source-file-structure
