+++ demos/3dr-solo-uav/3dr-solo-uav-react/src/components/SoloVideo.js
@@ -1,8 +1,8 @@
 import * as React from "react";
 import Draggable, {DraggableCore} from 'react-draggable'; // Both at the same time
 
-import VideoH264 from "osh/datareceiver/VideoH264";
-import FFMPEGView from "osh/ui/view/video/FFMPEGView";
+import VideoH264 from "osh/datareceiver/VideoH264.js";
+import FFMPEGView from "osh/ui/view/video/FFMPEGView.js";
 
 export class SoloVideoComponent extends React.Component {
   constructor(props) {
