+++ demos/3dr-solo-uav/3dr-solo-uav-vuejs/src/components/Video.vue
@@ -10,8 +10,8 @@
 
 <script>
 import DialogDrag from 'vue-dialog-drag';
-import VideoH264 from "osh/datareceiver/VideoH264";
-import FFMPEGView from "osh/ui/view/video/FFMPEGView";
+import VideoH264 from "osh/datareceiver/VideoH264.js";
+import FFMPEGView from "osh/ui/view/video/FFMPEGView.js";
 
 export default {
   name: "Video",
