+++ demos/video-display/video-display-vuejs/src/components/Map.vue
@@ -6,9 +6,9 @@
 
 <script>
     // @ is an alias to /src
-import LeafletView from "osh/ui/view/map/LeafletView";
-import SweJson from "osh/datareceiver/SweJson";
-import PointMarker from "osh/ui/styler/PointMarker";
+import LeafletView from "osh/ui/view/map/LeafletView.js";
+import SweJson from "osh/datareceiver/SweJson.js";
+import PointMarker from "osh/ui/styler/PointMarker.js";
 
 export default {
   name: "Map",
