+++ showcase/examples/cesium-fois/cesium-fois.js
@@ -1,7 +1,7 @@
-import CesiumView from 'osh/ui/view/map/CesiumView';
+import CesiumView from 'osh/ui/view/map/CesiumView.js';
 import {HeadingPitchRange, Math} from 'cesium';
-import PointMarker from 'osh/ui/styler/PointMarker';
-import Server from "osh/server/Server";
+import PointMarker from 'osh/ui/styler/PointMarker.js';
+import Server from "osh/server/Server.js";
 
 window.CESIUM_BASE_URL = './';
