+++ showcase/examples/leaflet-location-path/leaflet-location-path.js
@@ -2,7 +2,7 @@
 import SweJson from "osh/datareceiver/SweJson.js";
 import PointMarker from "osh/ui/styler/PointMarker.js";
 import LeafletView from "osh/ui/view/map/LeafletView.js";
-import Polyline from "osh/ui/styler/Polyline";
+import Polyline from "osh/ui/styler/Polyline.js";
 
 let gpsDataSource = new SweJson("android-GPS", {
     protocol: "ws",
