+++ showcase/examples/ptz-tasking/ptz-tasking.js
@@ -1,7 +1,7 @@
 import VideoH264 from 'osh/datareceiver/VideoH264.js';
 import FFMPEGView from 'osh/ui/view/video/FFMPEGView.js';
-import FoscamPtzTasking from "osh/datasender/FoscamPtzTasking";
-import PtzTaskingView from "osh-ext/ui/view/tasking/PtzTaskingView";
+import FoscamPtzTasking from "osh/datasender/FoscamPtzTasking.js";
+import PtzTaskingView from "osh-ext/ui/view/tasking/PtzTaskingView.js";
 
 // create data source for UAV camera
 let videoDataSource = new VideoH264("drone-Video", {
