+++ source/osh-ext/ui/view/tasking/PtzTaskingView.js
@@ -14,10 +14,10 @@
 
  ******************************* END LICENSE BLOCK ***************************/
 
-import View from "../../../../osh/ui/view/View";
-import {isDefined, randomUUID} from "../../../../osh/utils/Utils";
+import View from "../../../../osh/ui/view/View.js";
+import {isDefined, randomUUID} from "../../../../osh/utils/Utils.js";
 import "../../../resources/css/tasking.css";
-import EventManager from "../../../../osh/events/EventManager";
+import EventManager from "../../../../osh/events/EventManager.js";
 
 export const htmlTaskingComponent=`
         <div class=\"ptz-zoom\">
