+++ source/osh/OSH.js
@@ -1,33 +1,33 @@
 export const VERSION = '2.0.0';
 export { default as  Buffer } from './buffer/Buffer.js';
-export { default as  Ajax } from './dataconnector/Ajax';
-export { default as  DataConnector } from './dataconnector/DataConnector';
-export { default as  WebSocketConnector } from './dataconnector/WebSocketConnector';
-export { default as  DataReceiverController } from './datareceiver/DataReceiverController';
-export { default as  DataSource } from './datareceiver/DataSource';
-export { default as  SweJsonDataSource } from './datareceiver/SweJson';
-export { default as  OrientationQuaternion  } from '../osh-ext/datareceiver/OrientationQuaternion';
-export { default as  VideoH264 } from './datareceiver/VideoH264';
-export { default as  VideMjpeg } from './datareceiver/VideoMjpeg';
-export { default as  DataSenderController } from './datasender/DataSenderController';
-export { default as  DataSink } from './datasender/DataSink';
-export { default as  FoscamPtzTasking } from './datasender/FoscamPtzTasking';
-export { default as  PtzTasking } from './datasender/PtzTasking';
-export { default as  UavMapTasking } from './datasender/UavMapTasking';
-export { default as  EventManager  } from './events/EventManager';
-export { default as  EventMap } from './events/EventMap';
-export { default as  SWEXmlStreamParser  } from './parsers/SWEXmlStreamParser';
-export { default as  Server } from './server/Server';
-export { default as  Styler } from './ui/styler/Styler';
-export { default as  Curve } from './ui/styler/Curve';
-export { default as  ImageDraping } from './ui/styler/ImageDraping';
-export { default as  PointMarker } from './ui/styler/PointMarker';
-export { default as  Polyline } from './ui/styler/Polyline';
-export { default as  ChartJsView } from './ui/view/chart/ChartJsView';
-export { default as  EntityTreeView } from './ui/view/entity/EntityTreeView';
-export { default as  CesiumView } from './ui/view/map/CesiumView';
-export { default as  OpenLayerView } from './ui/view/map/OpenLayerView';
-export { default as  LeafletView  } from './ui/view/map/LeafletView';
-export { default as  FFMPEGView } from './ui/view/video/FFMPEGView';
-export { default as  MjpegView } from './ui/view/video/MjpegView';
-export { default as  View } from './ui/view/View';
+export { default as  Ajax } from './dataconnector/Ajax.js';
+export { default as  DataConnector } from './dataconnector/DataConnector.js';
+export { default as  WebSocketConnector } from './dataconnector/WebSocketConnector.js';
+export { default as  DataReceiverController } from './datareceiver/DataReceiverController.js';
+export { default as  DataSource } from './datareceiver/DataSource.js';
+export { default as  SweJsonDataSource } from './datareceiver/SweJson.js';
+export { default as  OrientationQuaternion  } from '../osh-ext/datareceiver/OrientationQuaternion.js';
+export { default as  VideoH264 } from './datareceiver/VideoH264.js';
+export { default as  VideMjpeg } from './datareceiver/VideoMjpeg.js';
+export { default as  DataSenderController } from './datasender/DataSenderController.js';
+export { default as  DataSink } from './datasender/DataSink.js';
+export { default as  FoscamPtzTasking } from './datasender/FoscamPtzTasking.js';
+export { default as  PtzTasking } from './datasender/PtzTasking.js';
+export { default as  UavMapTasking } from './datasender/UavMapTasking.js';
+export { default as  EventManager  } from './events/EventManager.js';
+export { default as  EventMap } from './events/EventMap.js';
+export { default as  SWEXmlStreamParser  } from './parsers/SWEXmlStreamParser.js';
+export { default as  Server } from './server/Server.js';
+export { default as  Styler } from './ui/styler/Styler.js';
+export { default as  Curve } from './ui/styler/Curve.js';
+export { default as  ImageDraping } from './ui/styler/ImageDraping.js';
+export { default as  PointMarker } from './ui/styler/PointMarker.js';
+export { default as  Polyline } from './ui/styler/Polyline.js';
+export { default as  ChartJsView } from './ui/view/chart/ChartJsView.js';
+export { default as  EntityTreeView } from './ui/view/entity/EntityTreeView.js';
+export { default as  CesiumView } from './ui/view/map/CesiumView.js';
+export { default as  OpenLayerView } from './ui/view/map/OpenLayerView.js';
+export { default as  LeafletView  } from './ui/view/map/LeafletView.js';
+export { default as  FFMPEGView } from './ui/view/video/FFMPEGView.js';
+export { default as  MjpegView } from './ui/view/video/MjpegView.js';
+export { default as  View } from './ui/view/View.js';
