+++ source/osh/events/EventMap.js
@@ -14,7 +14,7 @@
 
  ******************************* END LICENSE BLOCK ***************************/
 
-import {isDefined} from '../utils/Utils';
+import {isDefined} from '../utils/Utils.js';
 
 /**
  * This class is responsible for observing and firing events. It used an object as backed data structure.
